Started process of adding cards to the view schedules screen
