+++ frontend/frontend/lib/src/app.dart
@@ -4,6 +4,7 @@ import 'package:frontend/src/provider/google_sign_in.dart';
 import 'package:frontend/src/screens/Booking/bookingAdminHomeScreen.dart';
 import 'package:frontend/src/screens/Booking/bookingDayScreen.dart';
 import 'package:frontend/src/screens/Booking/bookingHomeScreen.dart';
+import 'package:frontend/src/screens/Booking/bookingScheduleScreen.dart';
 import 'package:frontend/src/screens/Login/googleAuthTest.dart';
 import 'package:frontend/src/screens/Noticeboard/noticeboardScreen.dart';
 import 'package:frontend/src/screens/Login/supplyInfoScreen.dart';
@@ -27,7 +28,7 @@ class AppState extends State<App> {
     return ChangeNotifierProvider(
       create: (context) => GoogleSignInProvider(),
       child: MaterialApp(
-        home: Login(),
+        home: BookingScreen(),
       ),
     );
   }
